+++ __init__.py
@@ -1,4 +1,4 @@
-VERSION = '0.7.4'
+VERSION = '1.0.0'
 
 import sys
 from os import path