+++ webview/src/api/server-api/networking/index.js
@@ -2,7 +2,7 @@ import { isDevMode } from "../../devMode";
 import { store } from "../../redux";
 import { setApiBaseUrl } from "../../redux/slices/apiBaseUrl";
 
-export let DEV_SERVER_URL = "https://www.dev.anki.rankmd.org";
+export let DEV_SERVER_URL = "https://dev.anki.rankmd.org";
 export let PROD_SERVER_URL = "https://anki.rankmd.org";
 
 export function setupServerAPI() {